+++ app/lib/contract-utils.ts
@@ -1,10 +1,12 @@
 import { type Address, type PublicClient } from "viem";
-import { NONCE_TRACKER_ADDRESS, ERC1967_IMPLEMENTATION_SLOT, MAGIC_PREFIX, PROXY_TEMPLATE_ADDRESSES, NEW_IMPLEMENTATION_ADDRESS, ENTRYPOINT_ADDRESS } from "./contracts";
+import { NONCE_TRACKER_ADDRESS, ERC1967_IMPLEMENTATION_SLOT, MAGIC_PREFIX, EIP7702PROXY_TEMPLATE_ADDRESS, CBSW_IMPLEMENTATION_ADDRESS, ENTRYPOINT_ADDRESS } from "./constants";
 import type { P256Credential } from "viem/account-abstraction";
 import { EntryPointAbi } from "./abi/EntryPoint";
 
 export const MIN_DEPOSIT = BigInt("100000000000000000"); // 0.1 ETH
 
+// Gets the next nonce for the account in the NonceTracker contract,
+// used for calls to `EIP7702Proxy.setImplementation`.
 export async function getNonceFromTracker(
   publicClient: PublicClient,
   account: Address
@@ -25,6 +27,7 @@ export async function getNonceFromTracker(
   return nonce;
 }
 
+// Gets the current implementation address from the ERC1967 slot at the given address
 export async function getCurrentImplementation(
   publicClient: PublicClient,
   address: Address
@@ -40,6 +43,7 @@ export async function getCurrentImplementation(
   return `0x${implementationSlotData.slice(-40)}` as Address;
 }
 
+// The current state of the contract at the given address
 export type ContractState = {
   bytecode: string;
   implementation: Address;
@@ -47,10 +51,12 @@ export type ContractState = {
   isImplementationDisrupted: boolean;
 };
 
+// The expected bytecode for a EOA upgraded to a CoinbaseSmartWallet, including the EIP-7702 magic prefix
 export function getExpectedBytecode() {
-  return `${MAGIC_PREFIX}${PROXY_TEMPLATE_ADDRESSES.odyssey.slice(2).toLowerCase()}`.toLowerCase();
+  return `${MAGIC_PREFIX}${EIP7702PROXY_TEMPLATE_ADDRESS.slice(2).toLowerCase()}`.toLowerCase();
 }
 
+// Checks the current state of the contract at the given address with respect to the expected bytecode and ERC1967 implementation address
 export async function checkContractState(
   publicClient: PublicClient,
   address: Address
@@ -67,7 +73,7 @@ export async function checkContractState(
   const isDelegateDisrupted = currentBytecode !== "0x" && currentBytecode !== expectedBytecode;
 
   // Check if implementation is disrupted
-  const isImplementationDisrupted = implementation.toLowerCase() !== NEW_IMPLEMENTATION_ADDRESS.toLowerCase();
+  const isImplementationDisrupted = implementation.toLowerCase() !== CBSW_IMPLEMENTATION_ADDRESS.toLowerCase();
 
   return {
     bytecode: bytecode || "0x",
@@ -77,6 +83,7 @@ export async function checkContractState(
   };
 }
 
+// Verifies that the given passkey is an owner of the given wallet address
 export async function verifyPasskeyOwnership(
   publicClient: PublicClient,
   walletAddress: Address,
@@ -104,12 +111,14 @@ export async function verifyPasskeyOwnership(
   return isOwner;
 }
 
+// The balances of the account at the given address, including the account balance and its current entrypoint deposit
 export type AccountBalances = {
   accountBalance: bigint;
   entryPointDeposit: bigint;
   needsDeposit: boolean;
 };
 
+// Checks the balances of the account at the given address, including the account balance and its current entrypoint deposit
 export async function checkAccountBalances(
   publicClient: PublicClient,
   address: Address
